% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohensd_OG.R
\name{cohensd_OG}
\alias{cohensd_OG}
\title{Generate a Cohen's d for two random normally distributed data sets}
\usage{
cohensd_OG(num, bef, dif, stdev)
}
\arguments{
\item{num}{number of subjects}

\item{bef}{mean}

\item{dif}{how much to add to the mean}

\item{stdev}{standard deviation}
}
\value{
Cohen's d
}
\description{
Generates a Cohen's d effect size for two random normally distributed data sets with defined characteristics (number of subjects, mean, difference between means, and standard deviation)
}
\examples{
cohensd_OG(75,0,1,1)
}
