% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohensdcalc.R
\name{cohensdcalc}
\alias{cohensdcalc}
\title{Generate a Cohen's d for two data sets}
\usage{
cohensdcalc(before, after)
}
\arguments{
\item{before}{data set of the before group}

\item{after}{data set of the after group}
}
\value{
Cohen's d
}
\description{
Generates a Cohen's d effect size for two data sets
}
\examples{
beforegroup<-rnorm(75,0,1)
aftergroup<-rnorm(75,1,1)
cohensdcalc(beforegroup,aftergroup)
}
