% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidint.R
\name{confidint}
\alias{confidint}
\title{Generate a confidence interval for a given x}
\usage{
confidint(x, num)
}
\arguments{
\item{x}{what to generate a confidence interval of}

\item{num}{number of subjects}
}
\value{
confidence interval
}
\description{
Generates a confidence interval for a given x and number of subjects in the data sets
}
\examples{
x <- replicate(100, cohensd_OG(75,0,1,1))
confidint(x,75)
}
