% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltamad.R
\name{deltamad}
\alias{deltamad}
\title{Generate a delta MAD for two random normally distributed data sets}
\usage{
deltamad(num, bef, dif, stdev)
}
\arguments{
\item{num}{number of subjects}

\item{bef}{mean}

\item{dif}{how much to add to the mean}

\item{stdev}{standard deviation}
}
\value{
Cohen's d
}
\description{
Generates a delta MAD effect size for two random normally distributed data sets with defined characteristics (number of subjects, mean, difference between means, and standard deviation)
}
\examples{
deltamad(75,0,1,1)
}
