% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltamad_A.R
\name{deltamad_A}
\alias{deltamad_A}
\title{Generate a delta MAD for one randomly distributed data set against a normal data set contaminated with another normal data set}
\usage{
deltamad_A(num0, num1, num2, bef, aft1, aft2, stdev0, stdev1, stdev2)
}
\arguments{
\item{num0}{number of subjects}

\item{num1}{percent/decimal of subjects for the first group}

\item{num2}{percent/decimal of subjects for the second group}

\item{bef}{mean of the original group}

\item{aft1}{mean for the first group}

\item{aft2}{mean for the second group}

\item{stdev0}{standard deviation of the original group}

\item{stdev1}{standard deviation for the first group}

\item{stdev2}{standard deviation mean for the second group}
}
\value{
Cohen's d
}
\description{
Generates a delta MAD effect size for one randomly distributed data set against a normal data set contaminated with another normal data set with defined characteristics (number of subjects, percent of subjects with x mean and y standard deviation, percent of subjects with z mean and a standard deviation, mean of the first data set, x mean, z mean, standard deviation of the second data set, y standard deviation, and a standard deviation)
}
\examples{
deltamad_A(75,0.8,0.2,0,1,0.5,1,1,2)
}
