% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltamadcalc.R
\name{deltamadcalc}
\alias{deltamadcalc}
\title{Generate a delta MAD for two data sets}
\usage{
deltamadcalc(num, before, after)
}
\arguments{
\item{num}{number of subjects}

\item{before}{data set of the before group}

\item{after}{data set of the after group}
}
\value{
delta MAD
}
\description{
Generates a delta MAD effect size for two data sets
}
\examples{
numberofsubjects<-75
beforegroup<-rnorm(numberofsubjects,0,1)
aftergroup<-rnorm(numberofsubjects,1,1)
deltamadcalc(numberofsubjects,beforegroup,aftergroup)
}
